requirements.txt
streamlit
faiss-cpu
openai
numpy
tiktoken
python-dotenv
requests
beautifulsoup4
pdfplumber
langdetect
google-genai
google-genai
